crawler/requirements.txt
Scrapy==1.7.4
google-cloud-language
pytz
